as6303: add initial bits so we can feed it Jcc instructions
The 6803/303 doesn't really have a long conditional branch, so like x86 we
need a smart branch that turns into either Bcc or B!cc 3 JMP addr according
to the distance. This needs some core assembler changes to do the optimization
so for now always generate the pessimal case.

In the longer term we need to do an extra pass which isn't too hard so that

pass 0: we compute worst cases
pass 1: we make decisions and compute real addresses (whilst tweaking the code
	that right now will throw a phase error)
pass 2: we generate the final code and output it.

We could do it in the optimizer when it's done, but given the fact it is a
useful asm feature it makes more sense to live here.