as68: correct TREL8 handling
It's possible for a forward TREL8 relocation (ie a Bcc instruction) to
appear to be out of range when the code before it shrinks dramatically
in pass 2. At the point the TREL8 is evaluated on pass 2 the actual symbol
value we have may be pre-shrink and thus give a misleading error.

Possibly we should just defer error processing to pass 3 but that has a perf
impact ?