ld: rework various things for absolute mode and more segments
At this point we appear to correctly parse and process all the link records
for the kernel binary. Not yet verified that we do them correctly but that
will be something to do as the compiler support routines are finished and
there is actually something to 'run' and knock all the compiler and link bugs
out with.