cc68: first attempt at fixing the endianness problems
For 6502 the cc65 compiler is little endian. Thus it takes some slightly
slack approaches to casting because it knows that an expression from
expr.c that is 'pointer to data is in the working register' can be resolved
with the type of the target as it will just load the correct low bytes.

On 680x with a big endian CPU loading the wrong type is fatal because if
you have something like

	int *x;
	char c = *x

you need to load 1,x not 0,x.

For 680x therefore do the evaluation in the true type then convert it. This
also means we can't sensibly use X for loads that will become char (because
we can't mask the top byte easily)