cc68: Fix string constant handling
This stops using the localstrings hack and uses the litpool properly. We can't
deal with references because our output path is more direct and lacks that
functionality.

Once things are a bit more stable I may look at switching it back to the
code used by cc65 to get this and some other things back common and better.