cc68: 6800 support remains
With these bits in place I can build fairly large amounts of pure 6800
code without the front end barfing or the assembler finding any invalid
operations.

The output is not great, it will need a much smarter optimizer than the
peephole hack we have now, and probably a lot more un-inlining.

The good thing is X is mostly free so we have a lot of opportunities for
inlines that do stuff like

	ldx #n
	jsr addxtod

	ldx #_sym
	jsr incptrby%d

a value/use tracking optimizer will also be able to make a much better dent
in the code output.