cc68: Change the 6800 ABI to differ from the others
The 6800 has no fast way to stack words via X and also has no nice way to clean
the stack up fast using pulx or abx patterns. Therefore switch the 6800 non
varargs ABI so that like 6502 the called function does the clean up.

It's not as pretty as 6502 because we don't have a separate C and CPU stack
but the ugly all hides in one support file so I guess that is OK.

This saves us a *lot* of space on call/return. There are also some optimization
rules to try and make the cases of pushing things more optimal particularly
when pushing several small constants.

We could take this further and make the 6800 like the 6502 fastcall ABI, but
this does require a core change to allow the code generator to remap the
offset of an argument (the last argument moves from 2,X to 0,X and return
from 0,x to 2,X)

With the work so far it's now possible to build (but not run!) a Fuzix kernel
on a pure 6800. It's still a bit tight in the 64K address space however.

The 6800 code generation still needs a fair bit of debugging to be useful