as68: allow for both small names for native and C99 names for cross build
There isn't much to note here except the weird S_SIZE bug

- Change the magic number for 32byte names as the formats are not compatible
- Set the assembler up to use 32byte names internally
- Use an extended version of the existing structures for 32 byte names
- Adjust the tools to print, access and the like correctly

Mixing formats is not supported (and not useful anyway)

The S_SIZE changes are a bit odd. S_SIZE was meant to be the size of a symbol
object. However it accidentally got used as the mask for the size bits. By
pure chance the size (19) was correct enough to work for 8bit targets because
it masked the lower size bit (the only one used) and the right shift then
removed the extra mistakenly included bits.

S_SIZE is now the mask, S_ENTRYSIZE the size of a symbol. Hopefully that avoids
any future confusion.