cc68: deal with cases we take the address of a register variable
Whilst &regvar is illegal in C there are cases we generate it internally
in the compiler. In that case we want to generate

	ldx #reg+n

not the nonsensicl

	ldx #@sreg+n

For 680x it rarely occurs but for 6800 it's actually quite common.

With this change the Fuzix libc passes the compile phase when building for
6800. Code correctness is still another matter altogether.
cc68: deal with cases we take the address of a register variable

Whilst &regvar is illegal in C there are cases we generate it internally
in the compiler. In that case we want to generate

	ldx #reg+n

not the nonsensicl

	ldx #@sreg+n

For 680x it rarely occurs but for 6800 it's actually quite common.

With this change the Fuzix libc passes the compile phase when building for
6800. Code correctness is still another matter altogether.
cc68: deal with cases we take the address of a register variable

Whilst &regvar is illegal in C there are cases we generate it internally
in the compiler. In that case we want to generate

	ldx #reg+n

not the nonsensicl

	ldx #@sreg+n

For 680x it rarely occurs but for 6800 it's actually quite common.

With this change the Fuzix libc passes the compile phase when building for
6800. Code correctness is still another matter altogether.
cc68: deal with cases we take the address of a register variable

Whilst &regvar is illegal in C there are cases we generate it internally
in the compiler. In that case we want to generate

	ldx #reg+n

not the nonsensicl

	ldx #@sreg+n

For 680x it rarely occurs but for 6800 it's actually quite common.

With this change the Fuzix libc passes the compile phase when building for
6800. Code correctness is still another matter altogether.
cc68: deal with cases we take the address of a register variable

Whilst &regvar is illegal in C there are cases we generate it internally
in the compiler. In that case we want to generate

	ldx #reg+n

not the nonsensicl

	ldx #@sreg+n

For 680x it rarely occurs but for 6800 it's actually quite common.

With this change the Fuzix libc passes the compile phase when building for
6800. Code correctness is still another matter altogether.
cc68: deal with cases we take the address of a register variable

Whilst &regvar is illegal in C there are cases we generate it internally
in the compiler. In that case we want to generate

	ldx #reg+n

not the nonsensicl

	ldx #@sreg+n

For 680x it rarely occurs but for 6800 it's actually quite common.

With this change the Fuzix libc passes the compile phase when building for
6800. Code correctness is still another matter altogether.
cc68: deal with cases we take the address of a register variable

Whilst &regvar is illegal in C there are cases we generate it internally
in the compiler. In that case we want to generate

	ldx #reg+n

not the nonsensicl

	ldx #@sreg+n

For 680x it rarely occurs but for 6800 it's actually quite common.

With this change the Fuzix libc passes the compile phase when building for
6800. Code correctness is still another matter altogether.
cc68: deal with cases we take the address of a register variable

Whilst &regvar is illegal in C there are cases we generate it internally
in the compiler. In that case we want to generate

	ldx #reg+n

not the nonsensicl

	ldx #@sreg+n

For 680x it rarely occurs but for 6800 it's actually quite common.

With this change the Fuzix libc passes the compile phase when building for
6800. Code correctness is still another matter altogether.
cc68: deal with cases we take the address of a register variable

Whilst &regvar is illegal in C there are cases we generate it internally
in the compiler. In that case we want to generate

	ldx #reg+n

not the nonsensicl

	ldx #@sreg+n

For 680x it rarely occurs but for 6800 it's actually quite common.

With this change the Fuzix libc passes the compile phase when building for
6800. Code correctness is still another matter altogether.