ld: Undo breakage with old CC68
The assembler and linker are shared with the Fuzix Compiler Kit. It does
a single pass on a sorted C library whilst the cc68 compiler is a cross
compiler doing multiple passes.

The change to the linker broke cc68. Default to no rescanning but allow
the cc68 build case to enable multiple passes.

Closes #23