6502: rework the branch auto-resizing
bcc and friends will now just error if they are out of range so code must
explicitly use "jcc" pseudo-ops for stuff that might be out of range and
should be auto-converted.

This allows for cases when the desired behaviour is a branch or error, and
matches how the other targets work.